+++ packages/eslint-plugin/babel.config.js
@@ -1,9 +1,14 @@
 module.exports = {
+  targets: "node 12.0",
   presets: [
-    ["@babel/preset-env", { targets: "node 12.0", modules: "commonjs" }],
+    ["@babel/preset-env", { modules: "commonjs" }],
     ["@babel/preset-typescript", { allowDeclareFields: true }],
   ],
   plugins: [
-    ["@babel/plugin-transform-runtime", { corejs: 3, version: "^7.17.9" }],
+    ["@babel/plugin-transform-runtime", { version: "^7.18.3" }],
+    [
+      "babel-plugin-polyfill-corejs3",
+      { method: "usage-pure", version: "^3.23.2" },
+    ],
   ],
 };