+++ packages/react/babel.config.js
@@ -1,9 +1,7 @@
 module.exports = {
+  targets: "defaults, not ie 11, not ie_mob 11",
   presets: [
-    [
-      "@babel/preset-env",
-      { targets: "defaults, not ie 11, not ie_mob 11", modules: "commonjs" },
-    ],
+    ["@babel/preset-env", { modules: "commonjs" }],
     // Use the classic runtime until we drop support for React < 18.0
     // because in older versions they don't have "exports" in package.json
     // and it does not work well with ESM.
@@ -12,6 +10,10 @@ module.exports = {
     ["@babel/preset-typescript", { allowDeclareFields: true }],
   ],
   plugins: [
-    ["@babel/plugin-transform-runtime", { corejs: 3, version: "^7.17.9" }],
+    ["@babel/plugin-transform-runtime", { version: "^7.18.3" }],
+    [
+      "babel-plugin-polyfill-corejs3",
+      { method: "usage-pure", version: "^3.23.2" },
+    ],
   ],
 };