+++ packages/react/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Add prepack script. It allows you to use unreleased versions from git with yarn v2 or later.
+
 ## 0.1.1
 
 - Add `Translate.Dynamic` for dynamically selecting translations