+++ packages/react/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add prepack script. It allows you to use unreleased versions from git with yarn v2 or later.
 - Add `"sideEffects": false` for better tree-shaking.
 - Add `useLocales` to get the context without immediately converting it to the translator object.
+  - `useI18n` now internally uses `useLocales`. Be careful if you have a locale named `""` (empty string), which I believe is very rare.
 
 ## 0.1.1